projects/python/upscale/pyproject.toml

[project]
name = "upscale"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = ["super-image", "opencv-python", "pillow"]
